make run_util.py for reproduce code easily
