+++ autorag/autorag/nodes/retrieval/run.py
@@ -2,14 +2,19 @@ import logging
 import os
 import pathlib
 from copy import deepcopy
-from typing import List, Callable, Dict, Tuple, Union
+from typing import List, Callable, Dict, Tuple
 
 import numpy as np
 import pandas as pd
 
-from autorag.evaluation import evaluate_retrieval
+from autorag.nodes.retrieval.run_util import (
+	run,
+	evaluate_retrieval_node,
+	save_and_summary,
+	find_best,
+)
 from autorag.schema.metricinput import MetricInput
-from autorag.strategy import measure_speed, filter_by_threshold, select_best
+from autorag.strategy import select_best
 from autorag.support import get_support_modules
 from autorag.utils.util import get_best_row, to_list, apply_recursive
 
@@ -59,107 +64,6 @@ def run_retrieval_node(
 	if not os.path.exists(save_dir):
 		os.makedirs(save_dir)
 
-	def run(input_modules, input_module_params) -> Tuple[List[pd.DataFrame], List]:
-		"""
-		Run input modules and parameters.
-
-		:param input_modules: Input modules
-		:param input_module_params: Input module parameters
-		:return: First, it returns list of result dataframe.
-		Second, it returns list of execution times.
-		"""
-		result, execution_times = zip(
-			*map(
-				lambda task: measure_speed(
-					task[0].run_evaluator,
-					project_dir=project_dir,
-					previous_result=previous_result,
-					**task[1],
-				),
-				zip(input_modules, input_module_params),
-			)
-		)
-		average_times = list(map(lambda x: x / len(result[0]), execution_times))
-
-		# run metrics before filtering
-		if strategies.get("metrics") is None:
-			raise ValueError("You must at least one metrics for retrieval evaluation.")
-		result = list(
-			map(
-				lambda x: evaluate_retrieval_node(
-					x,
-					metric_inputs,
-					strategies.get("metrics"),
-				),
-				result,
-			)
-		)
-
-		return result, average_times
-
-	def save_and_summary(
-		input_modules,
-		input_module_params,
-		result_list,
-		execution_time_list,
-		filename_start: int,
-	):
-		"""
-		Save the result and make summary file
-
-		:param input_modules: Input modules
-		:param input_module_params: Input module parameters
-		:param result_list: Result list
-		:param execution_time_list: Execution times
-		:param filename_start: The first filename to use
-		:return: First, it returns list of result dataframe.
-		Second, it returns list of execution times.
-		"""
-
-		# save results to folder
-		filepaths = list(
-			map(
-				lambda x: os.path.join(save_dir, f"{x}.parquet"),
-				range(filename_start, filename_start + len(input_modules)),
-			)
-		)
-		list(
-			map(
-				lambda x: x[0].to_parquet(x[1], index=False),
-				zip(result_list, filepaths),
-			)
-		)  # execute save to parquet
-		filename_list = list(map(lambda x: os.path.basename(x), filepaths))
-
-		summary_df = pd.DataFrame(
-			{
-				"filename": filename_list,
-				"module_name": list(map(lambda module: module.__name__, input_modules)),
-				"module_params": input_module_params,
-				"execution_time": execution_time_list,
-				**{
-					metric: list(map(lambda result: result[metric].mean(), result_list))
-					for metric in strategies.get("metrics")
-				},
-			}
-		)
-		summary_df.to_csv(os.path.join(save_dir, "summary.csv"), index=False)
-		return summary_df
-
-	def find_best(results, average_times, filenames):
-		# filter by strategies
-		if strategies.get("speed_threshold") is not None:
-			results, filenames = filter_by_threshold(
-				results, average_times, strategies["speed_threshold"], filenames
-			)
-		selected_result, selected_filename = select_best(
-			results,
-			strategies.get("metrics"),
-			filenames,
-			strategies.get("strategy", "mean"),
-		)
-		return selected_result, selected_filename
-
 	filename_first = 0
 	# run semantic modules
 	logger.info("Running retrieval node - semantic retrieval module...")
@@ -170,16 +74,28 @@ def run_retrieval_node(
 				zip(modules, module_params),
 			)
 		)
-		semantic_results, semantic_times = run(semantic_modules, semantic_module_params)
+		semantic_results, semantic_times = run(
+			semantic_modules,
+			semantic_module_params,
+			project_dir=project_dir,
+			previous_result=previous_result,
+			strategies=strategies,
+			metric_inputs=metric_inputs,
+		)
 		semantic_summary_df = save_and_summary(
 			semantic_modules,
 			semantic_module_params,
 			semantic_results,
 			semantic_times,
 			filename_first,
+			save_dir=save_dir,
+			strategies=strategies,
 		)
 		semantic_selected_result, semantic_selected_filename = find_best(
-			semantic_results, semantic_times, semantic_summary_df["filename"].tolist()
+			semantic_results,
+			semantic_times,
+			semantic_summary_df["filename"].tolist(),
+			strategies,
 		)
 		semantic_summary_df["is_best"] = (
 			semantic_summary_df["filename"] == semantic_selected_filename
@@ -201,16 +117,29 @@ def run_retrieval_node(
 				zip(modules, module_params),
 			)
 		)
-		lexical_results, lexical_times = run(lexical_modules, lexical_module_params)
+		lexical_results, lexical_times = run(
+			lexical_modules,
+			lexical_module_params,
+			project_dir=project_dir,
+			previous_result=previous_result,
+			strategies=strategies,
+			metric_inputs=metric_inputs,
+		)
+
 		lexical_summary_df = save_and_summary(
 			lexical_modules,
 			lexical_module_params,
 			lexical_results,
 			lexical_times,
 			filename_first,
+			save_dir=save_dir,
+			strategies=strategies,
 		)
 		lexical_selected_result, lexical_selected_filename = find_best(
-			lexical_results, lexical_times, lexical_summary_df["filename"].tolist()
+			lexical_results,
+			lexical_times,
+			lexical_summary_df["filename"].tolist(),
+			strategies,
 		)
 		lexical_summary_df["is_best"] = (
 			lexical_summary_df["filename"] == lexical_selected_filename
@@ -237,13 +166,22 @@ def run_retrieval_node(
 			["target_module_params" in x for x in hybrid_module_params]
 		):  # for Runner.run
 			# If target_module_params are already given, run hybrid retrieval directly
-			hybrid_results, hybrid_times = run(hybrid_modules, hybrid_module_params)
+			hybrid_results, hybrid_times = run(
+				hybrid_modules,
+				hybrid_module_params,
+				project_dir=project_dir,
+				previous_result=previous_result,
+				strategies=strategies,
+				metric_inputs=metric_inputs,
+			)
 			hybrid_summary_df = save_and_summary(
 				hybrid_modules,
 				hybrid_module_params,
 				hybrid_results,
 				hybrid_times,
 				filename_first,
+				save_dir=save_dir,
+				strategies=strategies,
 			)
 			filename_first += len(hybrid_modules)
 		else:  # for Evaluator
@@ -283,6 +221,8 @@ def run_retrieval_node(
 				hybrid_results,
 				hybrid_times,
 				filename_first,
+				save_dir=save_dir,
+				strategies=strategies,
 			)
 			filename_first += len(hybrid_modules)
 			hybrid_summary_df["execution_time"] = hybrid_times
@@ -324,7 +264,9 @@ def run_retrieval_node(
 	filenames = summary["filename"].tolist()
 
 	# filter by strategies
-	selected_result, selected_filename = find_best(results, average_times, filenames)
+	selected_result, selected_filename = find_best(
+		results, average_times, filenames, strategies
+	)
 	best_result = pd.concat([previous_result, selected_result], axis=1)
 
 	# add summary.csv 'is_best' column
@@ -339,35 +281,6 @@ def run_retrieval_node(
 	)
 	summary.to_csv(os.path.join(save_dir, "summary.csv"), index=False)
 	return best_result
-
-
-def evaluate_retrieval_node(
-	result_df: pd.DataFrame,
-	metric_inputs: List[MetricInput],
-	metrics: Union[List[str], List[Dict]],
-) -> pd.DataFrame:
-	"""
-	Evaluate retrieval node from retrieval node result dataframe.
-
-	:param result_df: The result dataframe from a retrieval node.
-	:param metric_inputs: List of metric input schema for AutoRAG.
-	:param metrics: Metric list from input strategies.
-	:return: Return result_df with metrics columns.
-	    The columns will be 'retrieved_contents', 'retrieved_ids', 'retrieve_scores', and metric names.
-	"""
-
-	@evaluate_retrieval(
-		metric_inputs=metric_inputs,
-		metrics=metrics,
-	)
-	def evaluate_this_module(df: pd.DataFrame):
-		return (
-			df["retrieved_contents"].tolist(),
-			df["retrieved_ids"].tolist(),
-			df["retrieve_scores"].tolist(),
-		)
-
-	return evaluate_this_module(result_df)
 
 
 def edit_summary_df_params(
